requirements.txt
Django
djangorestframework
psycopg2-binary
geopy
celery
redis
